trading-system/requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
PyYAML>=6.0.1

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data Sources
finnhub-python>=2.4.19
yfinance>=0.2.31
praw>=7.7.1
feedparser>=6.0.11
websocket-client>=1.6.4

# Sentiment / NLP
transformers>=4.35.0
torch>=2.1.0
anthropic>=0.39.0

# Technical Analysis
ta>=0.11.0

# Machine Learning
lightgbm>=4.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Execution
alpaca-py>=0.21.0

# Monitoring & Dashboard
streamlit>=1.28.0
plotly>=5.18.0
python-telegram-bot>=20.7

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Scheduling
apscheduler>=3.10.4

# Timezone handling
pytz>=2023.3
